pyproject.toml
[project]
name = "omniperf-bench"
version = "0.1.0"
description = "Benchmark framework for evaluating LLMs on software performance optimization tasks"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Kernel Corp"}
]
keywords = ["benchmark", "llm", "performance", "optimization", "vllm"]

dependencies = [
    "torch",
    "rich>=13.8.0",
    "pygithub>=2.4.0",
    "memory-profiler>=0.61.0",
    "fire>=0.6.0",
    "flask>=3.0.3",
    "fuzzywuzzy>=0.18.0",
    "python-levenshtein>=0.25.1",
    "seaborn>=0.13.2",
    "ghapi>=1.0.6",
    "scipy>=1.14.1",
    "datasets>=3.2.0",
    "huggingface-hub>=0.27.0",
    "docker>=7.1.0",
    "python-dotenv>=1.0.0",
    "boto3",
    "botocore",
    "pyyaml>=6",
    "jinja2>=3",
    "typer>=0.12",
    "pydantic>=2",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    # r2e: LLM completions library (required for commit analysis)
    "r2e @ git+https://github.com/r2e-project/r2e.git",
    "r2e-test-server>=0.1.4",
]

[project.optional-dependencies]
# Cloud GPU execution via Modal
modal = ["modal>=0.60.0"]

# Agent benchmarking (OpenHands integration)
agent-bench = [
    "openhands-ai>=0.10.0",
    "playwright>=1.45.0",
    "pytest-playwright>=0.4.2",
    "libtmux>=0.23.1",
]

# SkyPilot for distributed execution
skypilot = ["skypilot>=0.5.0"]

# Development tools
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

# LLM providers
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.20.0"]

# All optional dependencies
all = [
    "modal>=0.60.0",
    "openhands-ai>=0.10.0",
    "playwright>=1.45.0",
    "pytest-playwright>=0.4.2",
    "libtmux>=0.23.1",
    "skypilot>=0.5.0",
    "openai>=1.0.0",
    "anthropic>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]  # Line too long - handled by formatter

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
